+++ NEWS.md
@@ -1,6 +1,8 @@
 # spant 4.3.0
 * Bug fix for mr_data2bids function.
 * Bug fix for testing for GE data in fit_svs.
+* Added fit_svs_btrg_v1 function as a simple wrapper for fit_svs with defaults
+adjusted to improve the analysis of lower quality brain tumour spectra.
 
 # spant 4.2.0
 * Added rm_basis_elements function.